requirements.txt
ytmusicapi
flask
python-dotenv
dominate